+++ client/ayon_unreal/plugins/load/load_layout.py
@@ -200,7 +200,7 @@ class LayoutLoader(plugin.Loader):
         anim_file = Path(animation_file)
         anim_file_name = anim_file.with_suffix('')
 
-        anim_path = f"{asset_dir}/animations/{anim_file_name}"
+        anim_path = f"{asset_dir}/Animations/{anim_file_name}"
 
         folder_entity = get_current_folder_entity()
         # Import animation
@@ -832,7 +832,7 @@ class LayoutLoader(plugin.Loader):
         create_sequences = data["unreal"]["level_sequences_for_layouts"]
 
         root = "/Game/Ayon"
-        path = Path(container.get("namespace"))
+        path = Path(container["namespace"])
 
         containers = ls()
         layout_containers = [
@@ -945,16 +945,9 @@ class LayoutLoader(plugin.Loader):
             EditorLevelLibrary.load_level(tmp_level)
 
         # Delete the layout directory.
-        EditorAssetLibrary.delete_directory(str(path))
+        if EditorAssetLibrary.does_directory_exist(str(path)):
+            EditorAssetLibrary.delete_directory(str(path))
 
         if create_sequences:
             EditorLevelLibrary.load_level(master_level)
             EditorAssetLibrary.delete_directory(f"{root}/tmp")
-
-        # Delete the parent folder if there aren't any more layouts in it.
-        asset_content = EditorAssetLibrary.list_assets(
-            str(path.parent), recursive=False, include_folder=True
-        )
-
-        if len(asset_content) == 0:
-            EditorAssetLibrary.delete_directory(str(path.parent))
